web-scraper/requirements.txt
APScheduler==3.11.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Flask==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kafka-python==2.2.11
MarkupSafe==3.0.2
requests==2.32.3
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
Werkzeug==3.1.3
gunicorn